+++ CHANGELOG.md
@@ -16,6 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - False negative with nested imports in bad module attribute ([#30](https://github.com/duo-labs/dlint/issues/30))
 
+### Removed
+- `dlint.linters.helpers.bad_kwarg_use`: use of 'attribute_name' in favor of fully specified 'module_path' ([#19](https://github.com/duo-labs/dlint/issues/19))
+- `dlint.tree`: use of 'kwarg_attribute' in favor of 'kwarg_module_path' ([#21](https://github.com/duo-labs/dlint/issues/21))
+- `dlint.linters.helpers.bad_name_attribute_use`: use of attribute list in favor of fully specified module path ([#20](https://github.com/duo-labs/dlint/issues/20))
+
 ## [0.8.0] - 2019-09-18
 ### Added
 - Support for Python 3.8 ([#12](https://github.com/duo-labs/dlint/issues/12))